calculating velocity from ATL03; optimized reads of global ATL03 data

+++ plugin/GTArray.h
@@ -68,7 +68,7 @@ class GTArray
          * Methods
          *--------------------------------------------------------------------*/
 
-                    GTArray     (const Asset* asset, const char* resource, int track, const char* gt_dataset, H5Api::context_t* context, unsigned col=0, const long* prt_startrow=DefaultStartRow, const long* prt_numrows=DefaultNumRows);
+                    GTArray     (const Asset* asset, const char* resource, int track, const char* gt_dataset, H5Api::context_t* context, long col=0, const long* prt_startrow=DefaultStartRow, const long* prt_numrows=DefaultNumRows);
         virtual     ~GTArray    (void);
 
         bool        trim        (long* prt_offset);
@@ -98,7 +98,7 @@ const long GTArray<T>::DefaultNumRows[PAIR_TRACKS_PER_GROUND_TRACK] = {H5Api::AL
  * Constructor
  *----------------------------------------------------------------------------*/
 template <class T>
-GTArray<T>::GTArray(const Asset* asset, const char* resource, int track, const char* gt_dataset, H5Api::context_t* context, unsigned col, const long* prt_startrow, const long* prt_numrows):
+GTArray<T>::GTArray(const Asset* asset, const char* resource, int track, const char* gt_dataset, H5Api::context_t* context, long col, const long* prt_startrow, const long* prt_numrows):
     gt{ H5Array<T>(asset, resource, SafeString("/gt%dl/%s", track, gt_dataset).getString(), context, col, prt_startrow[PRT_LEFT], prt_numrows[PRT_LEFT]),
         H5Array<T>(asset, resource, SafeString("/gt%dr/%s", track, gt_dataset).getString(), context, col, prt_startrow[PRT_RIGHT], prt_numrows[PRT_RIGHT]) }
 {
